Try to salvage our use of the "Group" header into a semi consistent state for now.
The group "Development/Libraries/Haskell" is the only group that generated spec
files use by default. For a library package "foo", both "ghc-foo" and
"ghc-foo-devel" will be placed into that group. We don't use "System/Libraries"
for "ghc-foo", because [1] says that this group is inappropriate for the main
package and that system libraries should be placed into a sub-package. That
makes perfect sense, but it's a very extensive and difficult change because it
needs support from the build system provided by "ghc-rpm-macros".

Anyhow, these choices mean that executables written in Haskell will be placed
into that group, too, which is clearly not optimal. At some point in the
future, we'll have to clean this all up, but right now we can't.

[1] https://en.opensuse.org/openSUSE:Package_group_guidelines