cabal2obs: fail with an error if a package has missing dependencies
We cannot check the dependencies with great accuracy, i.e. we cannot verify
that the version number actually matches the constraints because we sometimes
add 'cabal-tweak-dep-ver' magic into the spec files. We would have to move that
information into the cabal2obs tool, too, so that it knows about overrides.