Revert all attempts to make Pandoc 3.x build.
It's way too difficult to get that done. We need the latest version of aeson
and vector and others to build pandoc, but those recent version conflict with
other packages all over the place. I suppose this is something we can do once
we're switching to ghc 9.x.

We're going back to revision 57ceeb6a18e37a36d15614c12899140ed7373ce8.