ghc-9.4.x: drop bytestring-builder from package set
The package is empty because its contents has been moved into the core
bytestring library that's part of GHC. Other builds should no longer depend on
it or use conditional logic if they do. We cannot build the empty package
because rpm-macros require a shared library to exist after the build, but the
empty package doesn't have one.