haskell-ci: we cannot build cabal2obs with ghc-9.10.x or ghc-9.12.x yet
Our code does not support the more recent versions of Cabal right now and
building an older version of Cabal with these newer compilers is tricky as
well. I'll make the necessary changes and update this project as well as
cabal2spec to the latest compiler version in the next few days.